Back to basics
Where the site is now hosted on S3, the 500 page is not really needed
(and there is no way to indicate when it would be shown). I am removing
all of the CSS, because it need to be "reworked" from the ground up. I
want a slim(er) site.

I will fix my style guide eventually.